refactor: split executor_login.go by concern
